change and use  presets for profiling/benchmarking
