+++ CMakeLists.txt
@@ -113,6 +113,7 @@ if(NEOFOAM_BUILD_TESTS)
 endif()
 
 if(NEOFOAM_BUILD_BENCHMARKS)
+  enable_testing()
   add_subdirectory(benchmarks)
 endif()
