+++ CMakePresets.json
@@ -37,7 +37,7 @@
         "CMAKE_BUILD_TYPE": "Debug",
         "NEOFOAM_DEVEL_TOOLS": "ON",
         "NEOFOAM_BUILD_TESTS": "ON",
-        "NEOFOAM_BUILD_BENCHMARKS": "ON",
+        "NEOFOAM_BUILD_BENCHMARKS": "OFF",
         "NEOFOAM_ENABLE_WARNINGS": "ON",
         "Kokkos_ENABLE_DEBUG": "ON",
         "Kokkos_ENABLE_DEBUG_BOUNDS_CHECK": "ON",
@@ -58,7 +58,7 @@
       "description": "Configuration for profiling use",
       "cacheVariables": {
         "CMAKE_BUILD_TYPE": "RelWithDebInfo",
-        "NEOFOAM_BUILD_TESTS": "ON",
+        "NEOFOAM_BUILD_TESTS": "OFF",
         "NEOFOAM_BUILD_BENCHMARKS": "ON",
         "NEOFOAM_ENABLE_WARNINGS": "OFF",
         "CMAKE_CXX_FLAGS": "-fno-omit-frame-pointer $env{CXXFLAGS}"
